feat: creates another project for accounts
