+++ src/ContoAuthApi.Accounts/Entities/WeatherForecast.cs
@@ -1,0 +1,6 @@
+namespace ContoAuthApi.Accounts.Entities;
+
+public record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
+{
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+}
