+++ src/ContoAuthApi.Accounts/Services/IWeatherForecastService.cs
@@ -1,0 +1,8 @@
+﻿using ContoAuthApi.Accounts.Entities;
+
+namespace ContoAuthApi.Accounts.Services;
+
+public interface IWeatherForecastService
+{
+    IEnumerable<WeatherForecast> GetWeatherForecasts();
+}
