+++ src/ContoAuthApi.Accounts/Services/WeatherForecastService.cs
@@ -1,0 +1,25 @@
+using ContoAuthApi.Accounts.Entities;
+
+namespace ContoAuthApi.Accounts.Services;
+
+public class WeatherForecastService : IWeatherForecastService
+{
+    public IEnumerable<WeatherForecast> GetWeatherForecasts()
+    {
+        var summaries = new[]
+        {
+            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        };
+
+        var forecast =  Enumerable.Range(1, 5).Select(index =>
+        new WeatherForecast
+        (
+            DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
+            Random.Shared.Next(-20, 55),
+            summaries[Random.Shared.Next(summaries.Length)]
+        ))
+        .ToArray();
+
+        return forecast;
+    }
+}
