+++ src/ContoAuthApi/ContoAuthApi.csproj
@@ -11,4 +11,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ContoAuthApi.Accounts\ContoAuthApi.Accounts.csproj" />
+  </ItemGroup>
+
 </Project>
